refactor: rename `with` module to `augment`
This also pulls retrieve logic into its own sub-module.
