+++ spnl/src/augment.rs
@@ -1,0 +1,8 @@
+pub mod embed;
+mod storage;
+
+mod index;
+pub use index::index;
+
+mod retrieve;
+pub use retrieve::retrieve;
