+++ spnl/src/augment/storage.rs
